+++ changelog.md
@@ -1,3 +1,8 @@
+### master
+
+- Rename `PhpVersionCompare` to `PhpVersionComparator` and move to `Steevanb\PhpVersionCompare\PhpVersion`
+- Add `assert()` methods for each major and minor PHP version
+
 ### 0.0.0 - 2021-10-23
 
 - Add `PhpVersionCompare`